code_wrapper: Created placeholder directory.
Created subpackage where all the dynamically-generated
code will reside.

generate_psa_wrappers.py has been updated to use the
new modules.
